added movies page
