+++ utils/urls.js
@@ -1,3 +1,3 @@
 const API_KEY = '963cfc8ae0b11289f364cb1de6574bcf'
 
-export const FETCH_URL = `https://api.themoviedb.org/3/movie/popular?api_key=${API_KEY}&language=en-US&page=1`
+export const FETCH_URL = `https://api.themoviedb.org/3/movie/top_rated?api_key=${API_KEY}&language=en-US&page=1`
